+++ lib/src/validators/pattern_validator.dart
@@ -8,11 +8,13 @@ import 'package:reactive_forms/src/validators/pattern/pattern_evaluator.dart';
 /// Validator that requires the control's value to match a regex pattern.
 class PatternValidator extends Validator<dynamic> {
   final PatternEvaluator evaluator;
+  final String validationMessage;
 
   /// Constructs an instance of [PatternValidator].
   ///
   /// The [evaluator] argument must not be null.
-  PatternValidator(this.evaluator);
+  PatternValidator(this.evaluator,
+      {this.validationMessage = ValidationMessage.pattern});
 
   @override
   Map<String, Object>? validate(AbstractControl<dynamic> control) {
@@ -21,7 +23,7 @@ class PatternValidator extends Validator<dynamic> {
             evaluator.hasMatch(control.value.toString()))
         ? null
         : {
-            ValidationMessage.pattern: {
+            validationMessage: {
               'requiredPattern': evaluator.pattern,
               'actualValue': control.value as Object,
             }